lambda/requirements.txt
aws-psycopg2==1.3.8
boto3==1.26.161
botocore==1.29.161
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
